another minor teak to js
The function to copy the text from the HernyB pic had an alert with a little bit in English regardless if it was the Portuguese or the English site. This tweak makes it so that the Portuguese site will have it all in Portuguese. I fixed that in the prior tweak, but then it was coping the intro to the clipboard and not just the alert. Fix that in this latest.